requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
groq==0.9.0
aiosqlite==0.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
jinja2==3.1.4
python-multipart==0.0.9